Add config and LED support
This commit adds configuration support to the program so that the source
code does not need to be edited to change the IoT Device primary key or
the time delay between program loop iterations. This makes it easier to
use the program in multiple environments and removes the risk of
accidentally committing secret keys to the repository.

RGB LED support has also been added to the program. Now any data read
from the accelerometer is written back to the LED so movement of the
board will be reflected in the color of the LED. The accelerometer
x-axis controls the red channel, y-axis controls the green channel, and
the z-axis controls the blue channel.

The program has also been slightly refactored so that log messages are
not intertwined with console output along with minor cleanups.