Add SIM7000A service files
This commit adds a systemd service file and associated scripts for
connecting and disconnecting from the cellular modem. With this service
properly installed the PocketBeagle can automatically start the network
on startup or simply provide a more convenient way to manually start and
stop the network. A configuration file for the Things Mobile network has
also been provided.